Major refactoring
Separated the loading in an initial loading and a new data loading. This makes it possible to just override the methods which deviate from the standard behaviour. Also, the control flow can be handled on a higher abstraction level, so we do not have to reproduce it many time.

Last, we added a field always_update to the Recorder to make it easier to unit test it.