[image] Added support for second camera.
Initially we just had a CCD camera which created xlsx files. Now we also have a CMOS camera which creates csv files. Moreover, they have different physical properties and thus the conversion factors are different.

To solve this issue, we separate the constants in two files and use dependency injection with a variable c (c_ccd or c_cmos) to be able to use the same code for two different cameras.

The default behaviour stays the same due to a trick in which we create a function instance that looks exactly like the previous function perform_analyses() but is in fact a method of a class instance.