[analysis] Major commit: Wrapper for SSDAnalysis.
Added a wrapper to the SSDAnalysis which can handle processing many files. This way, we only specify the folder and the time interval, and not the specific path of the ssd csv file.

This was necessary because we created many csv files instead of just one.

Moreover, I changed some methods from the ImageAnalysis to private such that we to not have to implement them in the Wrapper interface.