[analysis] Bug fix.
The following behaviour caused a bug: We called get_table() on the ImageParser which is a child of Record. In the case of no new data, the Parser returned the old data, which is again the specified behaviour.

To circumvent this, we first check if the Parser is up to data.

TODO: Fix behaviour of Parser instead.