[image] Dead pixel: Calculation and plotting.
Updated the calculation of the dead pixels and how they can be plotted at the beginning of the analysis.

TODO: Implement the following heuristic to find promising images:

Create a list of ‚real signal pixels‘ for each image based on a signal threshold and then check the number and position of the pixels.

If the pixels are
- numerous -> much signal
- clustered -> have a small variance in position, not by chance
then the image is more likely to contain real MOT signal.

Maybe this can be formalized with concepts like likelihood or entropy.