[online] Bug fix to make analysis run.
There was the problem that the recorder was up to date during the first run because the recorder was used to get the reference images for the fit_mot_number() method. Thus, the analysis was not performed by the Analysis parent class because it thought that it was already performed.